Revert "Rename ValidateTaskProperties to ValidatePlugins"
