+++ subprojects/plugin-development/src/main/java/org/gradle/plugin/devel/tasks/ValidateTaskProperties.java
@@ -16,14 +16,53 @@
 
 package org.gradle.plugin.devel.tasks;
 
-import org.gradle.api.DefaultTask;
+import com.google.common.base.Charsets;
+import com.google.common.base.Joiner;
+import com.google.common.collect.ImmutableList;
+import com.google.common.collect.Iterables;
+import com.google.common.collect.Lists;
+import com.google.common.collect.Maps;
+import com.google.common.io.Files;
+import org.gradle.api.GradleException;
+import org.gradle.api.InvalidUserDataException;
+import org.gradle.api.UncheckedIOException;
 import org.gradle.api.file.ConfigurableFileCollection;
+import org.gradle.api.file.EmptyFileVisitor;
+import org.gradle.api.file.FileVisitDetails;
 import org.gradle.api.file.RegularFileProperty;
-import org.gradle.api.tasks.Internal;
-import org.gradle.api.tasks.TaskProvider;
+import org.gradle.api.internal.ConventionTask;
+import org.gradle.api.internal.DocumentationRegistry;
+import org.gradle.api.model.ObjectFactory;
+import org.gradle.api.tasks.CacheableTask;
+import org.gradle.api.tasks.Classpath;
+import org.gradle.api.tasks.Input;
+import org.gradle.api.tasks.InputFiles;
+import org.gradle.api.tasks.Optional;
+import org.gradle.api.tasks.OutputFile;
+import org.gradle.api.tasks.PathSensitive;
+import org.gradle.api.tasks.PathSensitivity;
+import org.gradle.api.tasks.SkipWhenEmpty;
+import org.gradle.api.tasks.TaskAction;
+import org.gradle.api.tasks.TaskValidationException;
 import org.gradle.api.tasks.VerificationTask;
+import org.gradle.internal.classanalysis.AsmConstants;
+import org.gradle.internal.classloader.ClassLoaderFactory;
+import org.gradle.internal.classloader.ClassLoaderUtils;
+import org.gradle.internal.classpath.ClassPath;
+import org.gradle.internal.classpath.DefaultClassPath;
+import org.objectweb.asm.ClassReader;
+import org.objectweb.asm.ClassVisitor;
+import org.objectweb.asm.Opcodes;
 
 import javax.inject.Inject;
+import java.io.File;
+import java.io.IOException;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
 
 /**
  * Validates task property annotations.
@@ -63,25 +102,139 @@ import javax.inject.Inject;
  * </ul>
  *
  * @since 3.0
- *
- * @deprecated Use {@link ValidatePlugins} instead.
  */
-@Deprecated
-public class ValidateTaskProperties extends DefaultTask implements VerificationTask {
-    private final TaskProvider<ValidatePlugins> delegate;
-    private final Runnable deprecationNagger;
+@CacheableTask
+public class ValidateTaskProperties extends ConventionTask implements VerificationTask {
+    private final ConfigurableFileCollection classes;
+    private final ConfigurableFileCollection classpath;
+    private final RegularFileProperty outputFile;
+    private boolean enableStricterValidation;
+    private boolean ignoreFailures;
+    private boolean failOnWarning = true;
 
     @Inject
-    public ValidateTaskProperties(TaskProvider<ValidatePlugins> delegate, Runnable deprecationNagger) {
-        this.delegate = delegate;
-        this.deprecationNagger = deprecationNagger;
+    public ValidateTaskProperties(ObjectFactory objects) {
+        this.classes = objects.fileCollection();
+        this.classpath = objects.fileCollection();
+        this.outputFile = objects.fileProperty();
     }
 
-    @Internal
+    @TaskAction
+    public void validateTaskClasses() throws IOException {
+        ClassLoader previousContextClassLoader = Thread.currentThread().getContextClassLoader();
+        ClassPath classPath = DefaultClassPath.of(Iterables.concat(getClasses(), getClasspath()));
+        ClassLoader classLoader = getClassLoaderFactory().createIsolatedClassLoader("task-loader", classPath);
+        Thread.currentThread().setContextClassLoader(classLoader);
+        try {
+            validateTaskClasses(classLoader);
+        } finally {
+            Thread.currentThread().setContextClassLoader(previousContextClassLoader);
+            ClassLoaderUtils.tryClose(classLoader);
+        }
+    }
+
+    private void validateTaskClasses(final ClassLoader classLoader) throws IOException {
+        final Map<String, Boolean> taskValidationProblems = Maps.newTreeMap();
+        final Method validatorMethod;
+        try {
+            Class<?> validatorClass = classLoader.loadClass("org.gradle.api.internal.tasks.properties.PropertyValidationAccess");
+            validatorMethod = validatorClass.getMethod("collectValidationProblems", Class.class, Map.class, Boolean.TYPE);
+        } catch (ClassNotFoundException | NoSuchMethodException e) {
+            throw new RuntimeException(e);
+        }
+
+        getClasses().getAsFileTree().visit(new EmptyFileVisitor() {
+            @Override
+            public void visitFile(FileVisitDetails fileDetails) {
+                if (!fileDetails.getPath().endsWith(".class")) {
+                    return;
+                }
+                ClassReader reader;
+                try {
+                    reader = new ClassReader(Files.asByteSource(fileDetails.getFile()).read());
+                } catch (IOException e) {
+                    throw new UncheckedIOException(e);
+                }
+                List<String> classNames = Lists.newArrayList();
+                reader.accept(new TaskNameCollectorVisitor(classNames), ClassReader.SKIP_CODE);
+                for (String className : classNames) {
+                    Class<?> clazz;
+                    try {
+                        clazz = classLoader.loadClass(className);
+                    } catch (IllegalAccessError | NoClassDefFoundError | ClassNotFoundException e) {
+                        throw new GradleException("Could not load class: " + className, e);
+                    }
+                    try {
+                        validatorMethod.invoke(null, clazz, taskValidationProblems, enableStricterValidation);
+                    } catch (IllegalAccessException | InvocationTargetException e) {
+                        throw new RuntimeException(e);
+                    }
+                }
+            }
+        });
+        List<String> problemMessages = toProblemMessages(taskValidationProblems);
+        storeResults(problemMessages);
+        communicateResult(problemMessages, taskValidationProblems.containsValue(Boolean.TRUE));
+    }
+
+    private void storeResults(List<String> problemMessages) throws IOException {
+        if (outputFile.isPresent()) {
+            File output = outputFile.get().getAsFile();
+            //noinspection ResultOfMethodCallIgnored
+            output.createNewFile();
+            Files.asCharSink(output, Charsets.UTF_8).write(Joiner.on('\n').join(problemMessages));
+        }
+    }
+
+    private void communicateResult(List<String> problemMessages, boolean hasErrors) {
+        if (problemMessages.isEmpty()) {
+            getLogger().info("Task property validation finished without warnings.");
+        } else {
+            if (hasErrors || getFailOnWarning()) {
+                if (getIgnoreFailures()) {
+                    getLogger().warn("Task property validation finished with errors. See {} for more information on how to annotate task properties.{}", getDocumentationRegistry().getDocumentationFor("more_about_tasks", "sec:task_input_output_annotations"), toMessageList(problemMessages));
+                } else {
+                    throw new TaskValidationException(String.format("Task property validation failed. See %s for more information on how to annotate task properties.", getDocumentationRegistry().getDocumentationFor("more_about_tasks", "sec:task_input_output_annotations")), toExceptionList(problemMessages));
+                }
+            } else {
+                getLogger().warn("Task property validation finished with warnings:{}", toMessageList(problemMessages));
+            }
+        }
+    }
+
+    private static List<String> toProblemMessages(Map<String, Boolean> problems) {
+        ImmutableList.Builder<String> builder = ImmutableList.builder();
+        for (Map.Entry<String, Boolean> entry : problems.entrySet()) {
+            String problem = entry.getKey();
+            Boolean error = entry.getValue();
+            builder.add(String.format("%s: %s",
+                Boolean.TRUE.equals(error) ? "Error" : "Warning",
+                problem
+            ));
+        }
+        return builder.build();
+    }
+
+    private static CharSequence toMessageList(List<String> problemMessages) {
+        StringBuilder builder = new StringBuilder();
+        for (String problemMessage : problemMessages) {
+            builder.append(String.format("%n  - %s", problemMessage));
+        }
+        return builder;
+    }
+
+    private static List<InvalidUserDataException> toExceptionList(List<String> problemMessages) {
+        return problemMessages.stream()
+            .map(InvalidUserDataException::new)
+            .collect(Collectors.toList());
+    }
+
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public boolean getIgnoreFailures() {
-        deprecationNagger.run();
-        return delegate.get().getIgnoreFailures().get();
+        return ignoreFailures;
     }
 
     /**
@@ -89,8 +242,7 @@ public class ValidateTaskProperties extends DefaultTask implements VerificationT
      */
     @Override
     public void setIgnoreFailures(boolean ignoreFailures) {
-        deprecationNagger.run();
-        delegate.get().getIgnoreFailures().set(ignoreFailures);
+        this.ignoreFailures = ignoreFailures;
     }
 
     /**
@@ -98,28 +250,27 @@ public class ValidateTaskProperties extends DefaultTask implements VerificationT
      *
      * @since 4.0
      */
-    @Internal
+    @PathSensitive(PathSensitivity.RELATIVE)
+    @InputFiles
+    @SkipWhenEmpty
     public ConfigurableFileCollection getClasses() {
-        deprecationNagger.run();
-        return delegate.get().getClasses();
+        return classes;
     }
 
     /**
      * The classpath used to load the classes under validation.
      */
-    @Internal
+    @Classpath
     public ConfigurableFileCollection getClasspath() {
-        deprecationNagger.run();
-        return delegate.get().getClasspath();
+        return classpath;
     }
 
     /**
      * Returns whether the build should break when the verifications performed by this task detects a warning.
      */
-    @Internal
+    @Input
     public boolean getFailOnWarning() {
-        deprecationNagger.run();
-        return delegate.get().getFailOnWarning().get();
+        return failOnWarning;
     }
 
     /**
@@ -127,10 +278,9 @@ public class ValidateTaskProperties extends DefaultTask implements VerificationT
      *
      * @since 5.1
      */
-    @Internal
+    @Input
     public boolean getEnableStricterValidation() {
-        deprecationNagger.run();
-        return delegate.get().getEnableStricterValidation().get();
+        return enableStricterValidation;
     }
 
     /**
@@ -139,8 +289,7 @@ public class ValidateTaskProperties extends DefaultTask implements VerificationT
      * @since 5.1
      */
     public void setEnableStricterValidation(boolean enableStricterValidation) {
-        deprecationNagger.run();
-        delegate.get().getEnableStricterValidation().set(enableStricterValidation);
+        this.enableStricterValidation = enableStricterValidation;
     }
 
     /**
@@ -148,10 +297,10 @@ public class ValidateTaskProperties extends DefaultTask implements VerificationT
      *
      * @since 4.5
      */
-    @Internal
+    @Optional
+    @OutputFile
     public RegularFileProperty getOutputFile() {
-        deprecationNagger.run();
-        return delegate.get().getOutputFile();
+        return outputFile;
     }
 
     /**
@@ -159,8 +308,34 @@ public class ValidateTaskProperties extends DefaultTask implements VerificationT
      *
      * @param failOnWarning {@code true} to break the build on warning, {@code false} to ignore warnings. The default is {@code true}.
      */
+    @SuppressWarnings("unused")
     public void setFailOnWarning(boolean failOnWarning) {
-        deprecationNagger.run();
-        delegate.get().getFailOnWarning().set(failOnWarning);
+        this.failOnWarning = failOnWarning;
+    }
+
+    @Inject
+    protected ClassLoaderFactory getClassLoaderFactory() {
+        throw new UnsupportedOperationException();
+    }
+
+    @Inject
+    protected DocumentationRegistry getDocumentationRegistry() {
+        throw new UnsupportedOperationException();
+    }
+
+    private static class TaskNameCollectorVisitor extends ClassVisitor {
+        private final Collection<String> classNames;
+
+        public TaskNameCollectorVisitor(Collection<String> classNames) {
+            super(AsmConstants.ASM_LEVEL);
+            this.classNames = classNames;
+        }
+
+        @Override
+        public void visit(int version, int access, String name, String signature, String superName, String[] interfaces) {
+            if ((access & Opcodes.ACC_PUBLIC) != 0) {
+                classNames.add(name.replace('/', '.'));
+            }
+        }
     }
 }
